+++ Tests/LinuxMain.swift
@@ -1,0 +1,7 @@
+import XCTest
+@testable import EasyInject
+import EasyInjectTests
+
+var tests = [XCTestCaseEntry]()
+tests += EasyInjectTests.allTests()
+XCTMain(tests)